chore(quanta): clear pre-existing clippy warnings to satisfy -D warnings
Six small style fixes that the .githooks/pre-commit hook (cargo clippy
--workspace -- -D warnings) had been gating on:
- emit_constant_f16 marked #[allow(dead_code)] (intentional infra)
- collapsible-if let in emit_spirv/kernel.rs
- loop-index pattern in cpu/device.rs (×2)
- collapsible-if let in cpu/device.rs
- unused ToString import in cpu/exec.rs
- redundant u32-as-u32 cast in cpu/value.rs

No behavior change.

+++ src/driver/cpu/value.rs
@@ -196,7 +196,7 @@ pub(super) fn f16_to_f32(bits: u16) -> f32 {
                 e += 1;
             }
             f &= 0x3FF;
-            let f32_exp = (127 - 15 - e + 1) as u32;
+            let f32_exp = 127 - 15 - e + 1;
             f32::from_bits((sign << 31) | (f32_exp << 23) | (f << 13))
         }
     } else if exp == 31 {
